PJW/requirements.txt
pyyaml
requests
torchvision
uvicorn
fastapi
streamlit
pydantic
pillow